decorations: prevent interference from other lens extensions
VSCode has a long-standing bug
https://github.com/microsoft/vscode/issues/33852: multiple decorators on
the same line are resolved by whichever was inserted *last*. So you set
the **color** in the decoration **type** and the **contentText** in the
decoration **options**, another extension might insert its own style in
between.

This PR implements the suggested workaround: set the `contentText` and
`color` at the same time so they'll always appear together. It does so
setting everything in the decoration **options**, instead of only a
sparse `contentText`.

Related: https://github.com/usernamehw/vscode-error-lens/issues/25